Updates to missing part of the protocol + bugs
Updated the period mechanism so that the duty field represents a
fraction of the the period. e.g. duty = 100 is 100/255 of the period.
This a more "natural" behavipr

Updated the active flag to work more consistently. Current disbale
everything when 0, but nothing is reset. This reset functionalilty can
easily be commented back in.

Added the missing protocol elements for the vibe