Building automated player instatiation
Players are now being instantiated after the PlayerList dialog.
The players are only ever going to be created there. This player list is
then going to be passed to LevelSelect where player colors can be
modified from the default and then the player objects are passed to the
Level objects as they are instantiated