Somewhat hack build but functional
Update TODO to reflect findings

Serious issues with minimal packet sending as there is order loss or
packet loss resulting in essentially arbitray behavior. Currently I have
removed all the mechanism that prevent resending when the "new" value is
equal to the "old" value. Thus if you set the level to 255 when it is
already 255 it will resend nevertheless. This leads to huge amounts of
"excess" traffic however the result is quite usable since it curshes any
asyn/packet loss issues.

I have some ideas for solving this noted in todo, as well as some weird
bugs to resolve.

The Prox/Red LED issue is very serious on the patch and needs to be
isolated. As of now it's totally impossible to use red/prox channel.